Move client datatypes services to api/datatypes
