+++ client/src/api/datatypes.ts
@@ -1,8 +1,8 @@
-import { fetcher } from "@/api/schema/fetcher";
+import { fetcher } from "@/api/schema";
 
 const getTypesAndMappings = fetcher.path("/api/datatypes/types_and_mapping").method("get").create();
 
 export async function getDatatypes(upload_only = true) {
-    const response = await getTypesAndMappings({ upload_only });
-    return response.data;
+    const { data } = await getTypesAndMappings({ upload_only });
+    return data;
 }
