+++ client/src/components/Datatypes/factory.ts
@@ -1,5 +1,6 @@
+import { getDatatypes } from "@/api/datatypes";
+
 import { DatatypesMapperModel } from "./model";
-import { getDatatypes } from "./services";
 
 export async function getDatatypesMapper(upload_only = true) {
     const typesAndMapping = await getDatatypes(upload_only);
